backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
tensorflow==2.14.0
numpy==1.24.4
pillow==10.2.0
